issued_at is int, not string

+++ lib/OAuth/TokenIntrospection.php
@@ -89,8 +89,8 @@ class TokenIntrospection
         } else {
             // token exists and did not expire
             $r['active'] = TRUE;
-            $r['expires_at'] = $accessToken['issue_time'] + $accessToken['expires_in'];
-            $r['issued_at'] = $accessToken['issue_time'];
+            $r['expires_at'] = (int) ($accessToken['issue_time'] + $accessToken['expires_in']);
+            $r['issued_at'] = (int) $accessToken['issue_time'];
             $r['scope'] = $accessToken['scope'];
             $r['client_id'] = $accessToken['client_id'];
             $r['sub'] = $accessToken['resource_owner_id'];
